feat: 로또 티켓 및 당첨 티켓의 양방향 의존 관계 제거
        로또 티켓과 당첨 티켓이 양방향 의존 관계를 지니네요!
로또 티켓 자체는 당첨 티켓을 알 필요가 없어 보입니다 :)
당첨 티켓 스스로가 로또 티켓을 받아 어떤 등수인지 알려주게 하는 것은 어떨까요?

_Originally posted by @malibinYun in https://github.com/next-step/kotlin-lotto/pull/499#discussion_r1032990232_
