+++ src/main/kotlin/lotto/domain/WinningTicket.kt
@@ -4,11 +4,13 @@ class WinningTicket(
     private val winningTicket: LottoTicket,
     private val bonusNumber: LottoNumber
 ) {
-    fun containNumber(lottoNumber: LottoNumber): Boolean = winningTicket.containNumber(lottoNumber)
-    fun containBonusNumber(numbers: Set<LottoNumber>): Boolean = numbers.contains(bonusNumber)
-
     init {
         require(winningTicket.notContainNumber(bonusNumber)) { "$bonusNumber 은 이미 당첨 번호로 입력된 값이에요" }
+    }
+
+    fun match(lottoTicket: LottoTicket): Award {
+        val matchCount = lottoTicket.matchCount(winningTicket)
+        return Award.of(matchCount, lottoTicket.containNumber(bonusNumber))
     }
